chat noti in progress

+++ app/src/main/java/com/example/recruitment/model/Conversation.kt
@@ -1,9 +1,9 @@
 package com.example.recruitment.model
 
 data class Conversation(
-    var id: String = "", // Needed for document ID
+    var id: String = "",
     var participants: List<String> = listOf(),
     var lastMessage: String? = null,
     var lastTimestamp: Long? = 0L,
-    var unreadMessagesCount: Map<String, Int>? = null
+    var unreadMessagesCount: Map<String, Long>? = null
 )
